extract exporter tests into own test case
